+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- /contents view is sortable @ksuess
+
 ### Changes
 
 ## 4.1.2 (2020-03-30)